+++ react/UploadBlockWrapper.tsx
@@ -1,9 +1,10 @@
-import React from 'react'
+import React, { useState } from 'react'
 import { useQuery } from 'react-apollo'
 
 import ItemListContext from './ItemListContext'
 import GET_ACCOUNT_INFO from './queries/orderSoldToAccount.graphql'
 import UploadBlock from './UploadBlock'
+import { TableDataContext } from './utils/context'
 
 interface UploadBlockInterface {
   text?: string
@@ -18,6 +19,8 @@ const UploadBlockWrapper = ({
   componentOnly,
   downloadText,
 }: UploadBlockInterface) => {
+  const [tableData, setTableData] = useState<any>()
+
   const { ItemListProvider } = ItemListContext
 
   const { data: accountData, loading: accountDataLoading } = useQuery(
@@ -28,26 +31,47 @@ const UploadBlockWrapper = ({
     }
   )
 
+  const handleExtractData = (itemNumber, newData, dataType) => {
+    if (itemNumber === '-1') {
+      setTableData(newData)
+    } else {
+      setTableData((prevData) => {
+        return prevData?.map((item) => {
+          if (item.sku === itemNumber) {
+            return {
+              ...item,
+              [dataType]: newData,
+            }
+          } else {
+            return item
+          }
+        })
+      })
+    }
+  }
+
   return (
     <ItemListProvider
       accountData={accountData}
       accountDataLoading={accountDataLoading}
     >
-      <UploadBlock {...{ text, description, componentOnly, downloadText }} />
+      <TableDataContext.Provider value={{ tableData, handleExtractData }}>
+        <UploadBlock {...{ text, description, componentOnly, downloadText }} />
+      </TableDataContext.Provider>
     </ItemListProvider>
   )
 }
 
 UploadBlockWrapper.schema = {
-  "title": "Upload area inputs",
-  "type": "object",
-  "properties": {
-    "description": {
-      "type": "string",
-      "title": "description",
-      "description": "The label for the text input"
+  title: 'Upload area inputs',
+  type: 'object',
+  properties: {
+    description: {
+      type: 'string',
+      title: 'description',
+      description: 'The label for the text input',
     },
-  }
+  },
 }
 
 export default UploadBlockWrapper