+++ react/components/ItemPricing.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useState, useContext, useEffect } from 'react'
 import { useQuery } from 'react-apollo'
 import { useCssHandles } from 'vtex.css-handles'
 import getSymbolFromCurrency from 'currency-symbol-map'
@@ -7,6 +7,8 @@ import { useIntl, defineMessages } from 'react-intl'
 
 import GET_ITEM_PRICING from '../queries/getItemPricing.gql'
 import { getFormattedDate } from '../utils'
+import { TableDataContext } from '../utils/context'
+import type { TableData } from '../utils/context'
 
 import './ItemPricing.css'
 
@@ -53,8 +55,13 @@ const ItemPricing = ({ itemNumber, customerNumber, branch }: Props) => {
   const styles = useCssHandles(CSS_HANDLES)
   const [isOpen, setIsOpen] = useState(false)
   const intl = useIntl()
+  const { handleExtractData } = useContext(TableDataContext) as TableData
 
-  const { data: itemPricingInfo, loading } = useQuery(GET_ITEM_PRICING, {
+  const {
+    data: itemPricingInfo,
+    loading,
+    refetch,
+  } = useQuery(GET_ITEM_PRICING, {
     skip: !itemNumber || itemNumber === '',
     variables: {
       itemNumber,
@@ -74,6 +81,20 @@ const ItemPricing = ({ itemNumber, customerNumber, branch }: Props) => {
   const openModal = () => {
     setIsOpen(true)
   }
+
+  const refetchPriceListAndUpdateContext = async () => {
+    const { data } = await refetch()
+    const priceList = data?.getItemPricing?.itemPrices ?? []
+
+    handleExtractData(itemNumber, priceList, 'priceList')
+  }
+
+  useEffect(() => {
+    if (!loading) {
+      handleExtractData(itemNumber, priceList, 'priceList')
+    }
+    refetchPriceListAndUpdateContext()
+  }, [])
 
   return loading ? (
     <div className={`${styles.priceTable}`}>