+++ react/utils/buildURL.ts
@@ -1,0 +1,11 @@
+export const buildUrl = (
+  account: string,
+  acronym: string,
+  idWithouthAcronym: string,
+  fieldName: string,
+  file: string
+) => {
+  return account && acronym && idWithouthAcronym && fieldName && file
+    ? `https://${account}.vtexcommercestable.com.br/api/dataentities/${acronym}/documents/${idWithouthAcronym}/${fieldName}/attachments/${file}`
+    : ''
+}