+++ react/utils/const.ts
@@ -2,3 +2,130 @@ export const TARGET_SYSTEM = {
   JDE: 'JDE',
   SAP: 'SAP',
 }
+
+export const EMAIL_TEMPLATE_LOGO =
+  'https://sbdsefuat.vteximg.com.br/arquivos/email-temp-logo.png'
+
+export const LEGACY_SYSTEM_TABLE_SAP = [
+  {
+    header: 'STANLEY PART NUMBER',
+    key: 'skuName',
+    width: 30,
+  },
+  {
+    header: 'Description',
+    key: 'productName',
+    width: 40,
+  },
+  {
+    header: 'QTY PER UNIT',
+    key: 'uomDescription',
+    width: 15,
+  },
+  {
+    header: 'MOQ',
+    key: 'moq',
+    width: 15,
+  },
+  {
+    header: 'LEAD TIME',
+    key: 'leadTime',
+    width: 15,
+  },
+  // {
+  //   header: 'STOCKING UOM',
+  //   key: 'uom',
+  //   width: 15,
+  // },
+  {
+    header: 'Qty',
+    key: 'quantity',
+    width: 15,
+  },
+  // {
+  //   header: 'Price',
+  //   key: 'price',
+  //   width: 20,
+  // },
+  {
+    header: 'Avaliability',
+    key: 'availability',
+    width: 15,
+  },
+]
+
+export const LEGACY_SYSTEM_TABLE_JDE = [
+  {
+    header: 'STANLEY PART NUMBER',
+    key: 'skuName',
+    width: 30,
+  },
+  {
+    header: 'Description',
+    key: 'productName',
+    width: 40,
+  },
+  {
+    header: 'LEAD TIME',
+    key: 'leadTime',
+    width: 15,
+  },
+  {
+    header: 'STOCKING UOM',
+    key: 'uom',
+    width: 15,
+  },
+  {
+    header: 'PRICING UOM',
+    key: 'priceUom',
+    width: 15,
+  },
+  {
+    header: 'QTY PER UNIT',
+    key: 'uomDescription',
+    width: 15,
+  },
+  {
+    header: 'Weight',
+    key: 'weight',
+    width: 15,
+  },
+  {
+    header: 'TARIFF CODE',
+    key: 'tariffCode',
+    width: 15,
+  },
+  {
+    header: 'ORIGIN',
+    key: 'origin',
+    width: 15,
+  },
+  {
+    header: 'Qty',
+    key: 'quantity',
+    width: 15,
+  },
+  {
+    header: 'Price',
+    key: 'price',
+    width: 25,
+  },
+  {
+    header: 'Avaliability',
+    key: 'stockAvailability',
+    width: 15,
+  },
+]
+
+export const SAMPLE_TABLE = [
+  {
+    header: 'SKU',
+    key: 'SKU',
+    width: 15,
+  },
+  {
+    header: 'Quantity',
+    key: 'Quantity',
+    width: 15,
+  },
+]