+++ react/utils/context.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+export interface TableData {
+  tableData: any;
+  handleExtractData: (itemNumber: any, newData: any, dataType: string) => void;
+}
+
+export const TableDataContext = createContext<TableData | undefined>(undefined)