+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added thruDate as [DATE_TODAY + DATEADD(JDE_LEAD_TIME_DAYS)] (mm/dd/yyyy)
+- Added extra fields to quickorder table and excel report
 
 ## [0.1.80] - 2024-04-25